feature/cached-config-provider: Moved InvalidCacheKeyException to Exceptions subnamespace.
